backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
python-dotenv
openai
requests
pandas
openpyxl